pyproject.toml
[project]
name = "computer-vision-workshop"
version = "0.1.0"
description = "Workshop for NINA employees hosted by Ben and Taheera B)"
readme = "README.md"
authors = ["Taheera Ahmed <taheera.ahmed@nina.no>"]
requires-python = ">=3.10"
dependencies = [
    "torchvision>=0.20.1",
    "scikit-learn>=1.6.1",
    "lightning==2.4.0",
    "torch-tb-profiler>=0.4.3",
    "tabulate>=0.9.0",
    "gradio>=5.12.0",
    "hydra-core>=1.3.2",
    "ultralytics>=8.3.62",
    "albumentations>=2.0.0"
]
